# packages/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Agent Routing — AI Config API",
    "version": "0.1.0",
    "description": "Management API for AI agents and the knowledge (RAG) store. Consumed by the Next.js chat service via a generated TypeScript client."
  },
  "servers": [
    {
      "url": "/",
      "description": "Relative to the API host"
    }
  ],
  "paths": {
    "/agents": {
      "get": {
        "operationId": "listAgents",
        "summary": "List all agents (newest first)",
        "tags": [
          "agents"
        ],
        "responses": {
          "200": {
            "description": "Agents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createAgent",
        "summary": "Create a new agent",
        "tags": [
          "agents"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created agent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agents/{id}": {
      "get": {
        "operationId": "getAgent",
        "summary": "Fetch a single agent",
        "tags": [
          "agents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateAgent",
        "summary": "Update an agent (partial)",
        "tags": [
          "agents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated agent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteAgent",
        "summary": "Delete an agent",
        "tags": [
          "agents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/buckets": {
      "get": {
        "operationId": "listBuckets",
        "summary": "List knowledge buckets (with counts)",
        "tags": [
          "knowledge"
        ],
        "responses": {
          "200": {
            "description": "Buckets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BucketsResponse"
                }
              }
            }
          },
          "503": {
            "description": "RAG store unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createBucket",
        "summary": "Create a knowledge bucket",
        "tags": [
          "knowledge"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBucketInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created bucket",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BucketResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "RAG store unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/buckets/{id}": {
      "get": {
        "operationId": "getBucket",
        "summary": "Fetch a bucket plus its documents",
        "tags": [
          "knowledge"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bucket detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BucketDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "RAG store unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteBucket",
        "summary": "Delete a bucket (cascades documents + chunks)",
        "tags": [
          "knowledge"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "RAG store unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/buckets/{id}/documents": {
      "get": {
        "operationId": "listDocuments",
        "summary": "List a bucket's documents",
        "tags": [
          "knowledge"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Documents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentsResponse"
                }
              }
            }
          },
          "503": {
            "description": "RAG store unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "ingestDocument",
        "summary": "Ingest a document (chunk → embed → store)",
        "tags": [
          "knowledge"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestDocumentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ingested document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Bucket not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "RAG store unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/documents/{id}": {
      "delete": {
        "operationId": "deleteDocument",
        "summary": "Delete a document (cascades its chunks)",
        "tags": [
          "knowledge"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "RAG store unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/search": {
      "post": {
        "operationId": "searchKnowledge",
        "summary": "Run the full retrieval pipeline against one or more buckets",
        "tags": [
          "knowledge"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Retrieved chunks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "RAG store unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AgentInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "systemPrompt": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "temperature": {
            "type": "number"
          },
          "isRoutable": {
            "type": "boolean"
          },
          "isDefault": {
            "type": "boolean"
          },
          "builtinTools": {
            "type": "object",
            "properties": {
              "sendMessage": {
                "type": "boolean"
              },
              "deliverToAgent": {
                "type": "boolean"
              },
              "deliverToHuman": {
                "type": "boolean"
              },
              "endConversation": {
                "type": "boolean"
              }
            },
            "additionalProperties": false
          },
          "customTools": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                },
                "endpoint": {
                  "type": "string"
                }
              },
              "required": [
                "name",
                "description",
                "schema",
                "endpoint"
              ],
              "additionalProperties": false
            }
          },
          "mcpServers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "url": {
                  "type": "string"
                },
                "headers": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "name",
                "url"
              ],
              "additionalProperties": false
            }
          },
          "handoffRules": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "when": {
                  "type": "object",
                  "properties": {
                    "flag": {
                      "type": "string"
                    },
                    "keywords": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "additionalProperties": false
                },
                "assignTo": {
                  "type": "string"
                }
              },
              "required": [
                "when",
                "assignTo"
              ],
              "additionalProperties": false
            }
          },
          "guardrails": {
            "type": "object",
            "properties": {
              "enabled": {
                "type": "boolean"
              },
              "scope": {
                "type": "string"
              },
              "refusal": {
                "type": "string"
              }
            },
            "additionalProperties": false
          },
          "knowledge": {
            "type": "object",
            "properties": {
              "enabled": {
                "type": "boolean"
              },
              "bucketIds": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "topK": {
                "type": "number"
              }
            },
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
...
      "CreateBucketInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "provider": {
            "type": "string",
            "enum": [
              "local",
              "openai",
              "voyage"
            ]
          },
          "model": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "OkResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "additionalProperties": false
      },
...
      "AgentsResponse": {
        "type": "object",
        "properties": {
          "agents": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "systemPrompt": {
                  "type": "string"
                },
                "model": {
                  "type": "string"
                },
                "temperature": {
                  "type": "number"
                },
                "isRoutable": {
                  "type": "boolean"
                },
                "isDefault": {
                  "type": "boolean"
                },
                "builtinTools": {
                  "type": "object",
                  "properties": {
                    "sendMessage": {
                      "type": "boolean"
                    },
                    "deliverToAgent": {
                      "type": "boolean"
                    },
                    "deliverToHuman": {
                      "type": "boolean"
                    },
                    "endConversation": {
                      "type": "boolean"
                    }
                  },
                  "additionalProperties": false
                },
                "customTools": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "schema": {
                        "type": "object",
                        "additionalProperties": {}
                      },
                      "endpoint": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "name",
                      "description",
                      "schema",
                      "endpoint"
                    ],
                    "additionalProperties": false
                  }
                },
                "mcpServers": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "url": {
                        "type": "string"
                      },
                      "headers": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "string"
                        }
                      }
                    },
                    "required": [
                      "name",
                      "url"
                    ],
                    "additionalProperties": false
                  }
                },
                "handoffRules": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "when": {
                        "type": "object",
                        "properties": {
                          "flag": {
                            "type": "string"
                          },
                          "keywords": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        },
                        "additionalProperties": false
                      },
                      "assignTo": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "when",
                      "assignTo"
                    ],
                    "additionalProperties": false
                  }
                },
                "guardrails": {
                  "type": "object",
                  "properties": {
                    "enabled": {
                      "type": "boolean"
                    },
                    "scope": {
                      "type": "string"
                    },
                    "refusal": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false
                },
                "knowledge": {
                  "type": "object",
                  "properties": {
                    "enabled": {
                      "type": "boolean"
                    },
                    "bucketIds": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "topK": {
                      "type": "number"
                    }
                  },
                  "additionalProperties": false
                },
                "createdAt": {
                  "type": "string"
                },
                "updatedAt": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name",
                "description",
                "systemPrompt",
                "model",
                "temperature",
                "isRoutable",
                "isDefault",
                "builtinTools",
                "customTools",
                "mcpServers",
                "handoffRules",
                "guardrails",
                "knowledge",
                "createdAt",
                "updatedAt"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "agents"
        ],
        "additionalProperties": false
      },
...
      "BucketsResponse": {
        "type": "object",
        "properties": {
          "buckets": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "embeddingProvider": {
                  "type": "string",
                  "enum": [
                    "local",
                    "openai",
                    "voyage"
                  ]
                },
                "embeddingModel": {
                  "type": "string"
                },
                "embeddingDim": {
                  "type": "number"
                },
                "createdAt": {
                  "type": "string"
                },
                "documentCount": {
                  "type": "number"
                },
                "chunkCount": {
                  "type": "number"
                }
              },
              "required": [
                "id",
                "name",
                "description",
                "embeddingProvider",
                "embeddingModel",
                "embeddingDim",
                "createdAt"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "buckets"
        ],
        "additionalProperties": false
      },
...
    }
  }
}